+++ packages/contracts/contracts/CommitmentTreeManager.sol
@@ -60,11 +60,8 @@ contract CommitmentTreeManager is
         JoinSplit joinSplit
     );
 
-    /// @notice Event emitted when a new batch of notes is inserted into the tree
-    event InsertNotes(EncodedNote[] notes);
-
-    /// @notice Event emitted when a new batch of note commitments is inserted into the tree
-    event InsertNoteCommitments(uint256[] commitments);
+    /// @notice Event emitted when a subtree batch is filled with zeros
+    event FilledBatchWithZeros(uint256 startIndex, uint256 numZeros);
 
     /// @notice Event emitted when a subtree (and subsequently the main tree's root) are updated
     event SubtreeUpdate(uint256 newRoot, uint256 subtreeBatchOffset);
@@ -104,15 +101,12 @@ contract CommitmentTreeManager is
         uint256 batchLen = _merkle.getBatchLen();
         require(batchLen > 0, "!zero fill empty batch");
 
+        uint256 startIndex = _merkle.getTotalCount();
+        uint256 numZeros = TreeUtils.BATCH_SIZE - batchLen;
+
         _merkle.fillBatchWithZeros();
 
-        // instead of actually inserting the zeros, we emit an event saying we inserted zeros
-        // and then we call `_accumulate`. This prevents BATCH_SIZE - batchLen + 1 storage writes
-        uint256[] memory zeros = new uint256[](TreeUtils.BATCH_SIZE - batchLen);
-        for (uint256 i = 0; i < zeros.length; i++) {
-            zeros[i] = TreeUtils.ZERO_VALUE;
-        }
-        emit InsertNoteCommitments(zeros);
+        emit FilledBatchWithZeros(startIndex, numZeros);
     }
 
     /// @notice Attempts to update the tree's root given a subtree update proof
@@ -147,18 +141,16 @@ contract CommitmentTreeManager is
         return _merkle.getTotalCount();
     }
 
-    /// @notice Inserts a batch of notes into commitment tree
-    /// @param notes batch of notes to insert
-    function _insertNotes(EncodedNote[] memory notes) internal {
-        _merkle.insertNotes(notes);
-        emit InsertNotes(notes);
+    /// @notice Inserts note into commitment tree
+    /// @param note note to insert
+    function _insertNote(EncodedNote memory note) internal {
+        _merkle.insertNote(note);
     }
 
     /// @notice Inserts several note commitments into the tree
     /// @param ncs Note commitments to insert
     function _insertNoteCommitments(uint256[] memory ncs) internal {
         _merkle.insertNoteCommitments(ncs);
-        emit InsertNoteCommitments(ncs);
     }
 
     /// @notice Process several joinsplits, assuming that their proofs have already been verified
@@ -213,50 +205,34 @@ contract CommitmentTreeManager is
         _insertNoteCommitments(newNoteCommitments);
     }
 
-    /// @notice Inserts a batch of refund notes into the commitment tree
-    /// @param encodedAssets Encoded assets for each refund note
-    /// @param values Values for each refund note
-    /// @param refundAddr Stealth addresses for each refund note
-    /// @param numRefunds Number of refund notes to insert
-    function _handleRefundNotes(
-        EncodedAsset[] memory encodedAssets,
-        uint256[] memory values,
+    /// @notice Inserts a single refund note into the commitment tree
+    /// @param encodedAsset Encoded asset refund note is being created for
+    /// @param refundAddr Stealth address refund note is created to
+    /// @param value Value of refund note for given asset
+    function _handleRefundNote(
+        EncodedAsset memory encodedAsset,
         StealthAddress calldata refundAddr,
-        uint256 numRefunds
+        uint256 value
     ) internal {
-        uint256 numEncodedAssets = encodedAssets.length;
-        require(
-            numEncodedAssets == values.length && numRefunds <= numEncodedAssets,
-            "len mismatch"
+        uint128 index = _merkle.getTotalCount();
+        EncodedNote memory note = EncodedNote({
+            ownerH1: refundAddr.h1X,
+            ownerH2: refundAddr.h2X,
+            nonce: index,
+            encodedAssetAddr: encodedAsset.encodedAssetAddr,
+            encodedAssetId: encodedAsset.encodedAssetId,
+            value: value
+        });
+
+        _insertNote(note);
+
+        emit RefundProcessed(
+            refundAddr,
+            index,
+            encodedAsset.encodedAssetAddr,
+            encodedAsset.encodedAssetId,
+            value,
+            index
         );
-
-        if (numRefunds > 0) {
-            EncodedNote[] memory notes = new EncodedNote[](numRefunds);
-            uint128 offset = _merkle.getTotalCount();
-            for (uint256 i = 0; i < numRefunds; i++) {
-                EncodedAsset memory encodedAsset = encodedAssets[i];
-                uint256 value = values[i];
-
-                notes[i] = EncodedNote({
-                    ownerH1: refundAddr.h1X,
-                    ownerH2: refundAddr.h2X,
-                    nonce: offset + i,
-                    encodedAssetAddr: encodedAsset.encodedAssetAddr,
-                    encodedAssetId: encodedAsset.encodedAssetId,
-                    value: value
-                });
-
-                emit RefundProcessed(
-                    refundAddr,
-                    offset + i,
-                    encodedAsset.encodedAssetAddr,
-                    encodedAsset.encodedAssetId,
-                    value,
-                    offset + uint128(i)
-                );
-            }
-
-            _insertNotes(notes);
-        }
     }
 }