+++ packages/contracts/contracts/test/harnesses/TestBalanceManager.sol
@@ -25,13 +25,11 @@ contract TestBalanceManager is IHandler, BalanceManager {
     function handleDeposit(
         Deposit calldata deposit
     ) external override onlyTeller {
-        EncodedAsset[] memory encodedAssets = new EncodedAsset[](1);
-        uint256[] memory values = new uint256[](1);
-
-        encodedAssets[0] = deposit.encodedAsset;
-        values[0] = deposit.value;
-
-        _handleRefundNotes(encodedAssets, values, deposit.depositAddr, 1);
+        _handleRefundNote(
+            deposit.encodedAsset,
+            deposit.depositAddr,
+            deposit.value
+        );
     }
 
     function handleOperation(