+++ packages/sdk/src/syncSDK.ts
@@ -60,12 +60,15 @@ export async function syncSDK(
     console.log("applied state diff for block", diff.blockNumber);
 
     // TODO: deal with case where we have failure between applying state diff to DB and merkle being persisted
-    await updateMerkle(
-      merkle,
-      diff.lastCommittedMerkleIndex,
-      diff.notesAndCommitments.map(({ inner }) => inner),
-      nfIndices
-    );
+
+    if (diff.lastCommittedMerkleIndex) {
+      await updateMerkle(
+        merkle,
+        diff.lastCommittedMerkleIndex,
+        diff.notesAndCommitments.map(({ inner }) => inner),
+        nfIndices
+      );
+    }
   }
 }
 
@@ -100,6 +103,7 @@ async function updateMerkle(
   }
 
   // commit up to latest subtree commit
+  console.log("committing up to index", latestCommittedMerkleIndex);
   merkle.commitUpToIndex(latestCommittedMerkleIndex);
 
   console.log("merkle root:", merkle.getRoot());